app/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
croniter==6.0.0
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
openai==2.24.0
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyMuPDF==1.25.5
python-dotenv==1.2.2
pytz==2026.1.post1
redis==7.2.1
rq==2.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
waitress==3.0.2
Werkzeug==3.1.6
zope.event==6.1
zope.interface==8.2
